+++ allennlp_models/pair_classification/models/decomposable_attention.py
@@ -200,4 +200,4 @@ class DecomposableAttention(Model):
     def get_metrics(self, reset: bool = False) -> Dict[str, float]:
         return {"accuracy": self._accuracy.get_metric(reset)}
 
-    default_predictor = "textual-entailment"
+    default_predictor = "textual_entailment"