+++ allennlp_models/pair_classification/models/esim.py
@@ -240,3 +240,5 @@ class ESIM(Model):
 
     def get_metrics(self, reset: bool = False) -> Dict[str, float]:
         return {"accuracy": self._accuracy.get_metric(reset)}
+
+    default_predictor = "textual_entailment"