+++ allennlp_models/pretrained.py
@@ -1,5 +1,7 @@
+from typing import Optional
+
 from allennlp.models import load_archive
-from allennlp.predictors import Predictor, SentenceTaggerPredictor
+from allennlp.predictors import Predictor, TextClassifierPredictor
 
 from allennlp_models.coref import CorefPredictor
 from allennlp_models.pair_classification import DecomposableAttentionPredictor
@@ -10,11 +12,13 @@ from allennlp_models.structured_prediction import (
     ConstituencyParserPredictor,
     BiaffineDependencyParserPredictor,
 )
+from allennlp_models.tagging.predictors import SentenceTaggerPredictor
+
 
 # flake8: noqa: E501
 
 
-def _load_predictor(archive_file: str, predictor_name: str) -> Predictor:
+def _load_predictor(archive_file: str, predictor_name: Optional[str] = None) -> Predictor:
     """
     Helper to load the desired predictor from the given archive.
     """
@@ -24,8 +28,7 @@ def _load_predictor(archive_file: str, predictor_name: str) -> Predictor:
 
 def bert_srl_shi_2019() -> SemanticRoleLabelerPredictor:
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/bert-base-srl-2020.03.24.tar.gz",
-        "semantic-role-labeling",
+        "https://storage.googleapis.com/allennlp-public-models/bert-base-srl-2020.03.24.tar.gz"
     )
     return predictor
 
@@ -37,24 +40,21 @@ def bidirectional_attention_flow_seo_2017() -> ReadingComprehensionPredictor:
     Based on `BiDAF (Seo et al, 2017) <https://www.semanticscholar.org/paper/Bidirectional-Attention-Flow-for-Machine-Comprehen-Seo-Kembhavi/007ab5528b3bd310a80d553cccad4b78dc496b02>`_
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/bidaf-model-2020.03.19.tar.gz",
-        "reading-comprehension",
+        "https://storage.googleapis.com/allennlp-public-models/bidaf-model-2020.03.19.tar.gz"
     )
     return predictor
 
 
 def naqanet_dua_2019() -> ReadingComprehensionPredictor:
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/naqanet-2020.02.19.tar.gz",
-        "reading-comprehension",
+        "https://storage.googleapis.com/allennlp-public-models/naqanet-2020.02.19.tar.gz"
     )
     return predictor
 
 
 def open_information_extraction_stanovsky_2018() -> OpenIePredictor:
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/openie-model.2020.03.26.tar.gz",
-        "open-information-extraction",
+        "https://storage.googleapis.com/allennlp-public-models/openie-model.2020.03.26.tar.gz"
     )
     return predictor
 
@@ -66,8 +66,7 @@ def decomposable_attention_with_elmo_parikh_2017() -> DecomposableAttentionPredi
     Based on `Parikh et al, 2017 <https://www.semanticscholar.org/paper/A-Decomposable-Attention-Model-for-Natural-Languag-Parikh-T%C3%A4ckstr%C3%B6m/07a9478e87a8304fc3267fa16e83e9f3bbd98b27>`_
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/decomposable-attention-elmo-2020.04.09.tar.gz",
-        "textual-entailment",
+        "https://storage.googleapis.com/allennlp-public-models/decomposable-attention-elmo-2020.04.09.tar.gz"
     )
     return predictor
 
@@ -77,8 +76,7 @@ def neural_coreference_resolution() -> CorefPredictor:
     Coreference Resolution
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/coref-spanbert-large-2020.02.27.tar.gz",
-        "coreference-resolution",
+        "https://storage.googleapis.com/allennlp-public-models/coref-spanbert-large-2020.02.27.tar.gz"
     )
     return predictor
 
@@ -90,8 +88,7 @@ def named_entity_recognition_with_elmo_peters_2018() -> SentenceTaggerPredictor:
     Based on `Deep contextualized word representations <https://arxiv.org/abs/1802.05365>`_
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/ner-model-2020.02.10.tar.gz",
-        "sentence-tagger",
+        "https://storage.googleapis.com/allennlp-public-models/ner-model-2020.02.10.tar.gz"
     )
     return predictor
 
@@ -101,8 +98,7 @@ def fine_grained_named_entity_recognition_with_elmo_peters_2018() -> SentenceTag
     Fine Grained Named Entity Recognition
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/fine-grained-ner-model-elmo-2018.12.21.tar.gz",
-        "sentence-tagger",
+        "https://storage.googleapis.com/allennlp-public-models/fine-grained-ner-model-elmo-2018.12.21.tar.gz"
     )
     return predictor
 
@@ -114,8 +110,7 @@ def span_based_constituency_parsing_with_elmo_joshi_2018() -> ConstituencyParser
     Based on `Minimal Span Based Constituency Parser (Stern et al, 2017) <https://www.semanticscholar.org/paper/A-Minimal-Span-Based-Neural-Constituency-Parser-Stern-Andreas/593e4e749bd2dbcaf8dc25298d830b41d435e435>`_ but with ELMo embeddings
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/elmo-constituency-parser-2020.02.10.tar.gz",
-        "constituency-parser",
+        "https://storage.googleapis.com/allennlp-public-models/elmo-constituency-parser-2020.02.10.tar.gz"
     )
     return predictor
 
@@ -127,8 +122,7 @@ def biaffine_parser_universal_dependencies_todzat_2017() -> BiaffineDependencyPa
     Based on `Dozat and Manning, 2017 <https://arxiv.org/pdf/1611.01734.pdf>`_
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/biaffine-dependency-parser-ptb-2020.04.06.tar.gz",
-        "biaffine-dependency-parser",
+        "https://storage.googleapis.com/allennlp-public-models/biaffine-dependency-parser-ptb-2020.04.06.tar.gz"
     )
     return predictor
 
@@ -140,7 +134,32 @@ def esim_nli_with_elmo_chen_2017() -> DecomposableAttentionPredictor:
     Based on `Enhanced LSTM for Natural Language Inference <https://arxiv.org/pdf/1609.06038.pdf>`_ and uses ELMo
     """
     predictor = _load_predictor(
-        "https://storage.googleapis.com/allennlp-public-models/decomposable-attention-elmo-2020.04.09.tar.gz",
-        "textual-entailment",
+        "https://storage.googleapis.com/allennlp-public-models/decomposable-attention-elmo-2020.04.09.tar.gz"
     )
     return predictor
+
+
+def glove_sst() -> TextClassifierPredictor:
+    return _load_predictor(
+        "https://storage.googleapis.com/allennlp-public-models/basic_stanford_sentiment_treebank-2020.06.09.tar.gz"
+    )
+
+
+def roberta_sst() -> TextClassifierPredictor:
+    return _load_predictor(
+        "https://storage.googleapis.com/allennlp-public-models/sst-roberta-large-2020.06.08.tar.gz"
+    )
+
+
+def roberta_mnli() -> DecomposableAttentionPredictor:
+    return _load_predictor(
+        "https://storage.googleapis.com/allennlp-public-models/mnli_roberta-2020.06.09.tar.gz",
+        "textual-entailment",
+    )
+
+
+def roberta_snli() -> DecomposableAttentionPredictor:
+    return _load_predictor(
+        "https://storage.googleapis.com/allennlp-public-models/snli_roberta-2020.06.09.tar.gz",
+        "textual-entailment",
+    )