+++ allennlp_models/rc/predictors/bidaf.py
@@ -1,4 +1,3 @@
-from copy import deepcopy
 from typing import Dict, List
 
 from overrides import overrides
@@ -17,7 +16,7 @@ from allennlp.data.fields import (
 )
 
 
-@Predictor.register("reading-comprehension")
+@Predictor.register("reading_comprehension")
 class ReadingComprehensionPredictor(Predictor):
     """
     Predictor for the :class:`~allennlp_rc.models.bidaf.BidirectionalAttentionFlow` model, and any
@@ -57,7 +56,7 @@ class ReadingComprehensionPredictor(Predictor):
     def predictions_to_labeled_instances(
         self, instance: Instance, outputs: Dict[str, numpy.ndarray]
     ) -> List[Instance]:
-        new_instance = deepcopy(instance)
+        new_instance = instance.duplicate()
         # For BiDAF
         if "best_span" in outputs:
             span_start_label = outputs["best_span"][0]