+++ tests/coref/predictors/coref_test.py
@@ -16,7 +16,7 @@ class TestCorefPredictor(AllenNlpTestCase):
             "contains coreferent parts."
         }
         archive = load_archive(FIXTURES_ROOT / "coref" / "serialization" / "model.tar.gz")
-        predictor = Predictor.from_archive(archive, "coreference-resolution")
+        predictor = Predictor.from_archive(archive, "coreference_resolution")
 
         result = predictor.predict_json(inputs)
         self.assert_predict_result(result)
@@ -82,7 +82,7 @@ class TestCorefPredictor(AllenNlpTestCase):
 
         document = "This is a test sentence."
         archive = load_archive(FIXTURES_ROOT / "coref" / "serialization" / "model.tar.gz")
-        predictor = Predictor.from_archive(archive, "coreference-resolution")
+        predictor = Predictor.from_archive(archive, "coreference_resolution")
         result = predictor.coref_resolved(document)
         assert isinstance(result, str)
 
@@ -133,7 +133,7 @@ class TestCorefPredictor(AllenNlpTestCase):
             "contains coreferent parts."
         }
         archive = load_archive(FIXTURES_ROOT / "coref" / "serialization" / "model.tar.gz")
-        predictor = Predictor.from_archive(archive, "coreference-resolution")
+        predictor = Predictor.from_archive(archive, "coreference_resolution")
 
         instance = predictor._json_to_instance(inputs)
         outputs = predictor._model.forward_on_instance(instance)