+++ CHANGELOG.md
@@ -7,17 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.0.0](https://github.com/allenai/allennlp-models/releases/tag/v1.0.0) - 2020-06-16
+
+No additional note-worthy changes since rc6.
+
+## [v1.0.0rc6](https://github.com/allenai/allennlp-models/releases/tag/v1.0.0rc6) - 2020-06-11
+
 ### Changed
 
-- Nothing yet
+- Removed deprecated `"simple_seq2seq"` predictor
 
 ### Fixed
 
-- Nothing yet
+- Replaced `deepcopy` of `Instance`s with new `Instance.duplicate()` method.
+- A bug where pretrained sentence taggers would fail to be initialized because some of the models
+  were not imported.
+- A bug in some RC models that would cause mixed precision training to crash when using NVIDIA apex.
+- Predictor names were inconsistently switching between dashes and underscores. Now they all use underscores.
 
 ### Added
 
-- Nothing yet
+- Added option to SemanticDependenciesDatasetReader to not skip instances that have no arcs, for validation data
+- Added a default predictors to several models
+- Added sentiment analysis models to pretrained.py
+- Added NLI models to pretrained.py
 
 ## [v1.0.0rc5](https://github.com/allenai/allennlp-models/releases/tag/v1.0.0rc5) - 2020-05-14
 
@@ -27,7 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Made `transformer_qa` predictor accept JSON input with the keys "question" and "passage" to be consistent with the `reading-comprehension` predictor.
+- Made `transformer_qa` predictor accept JSON input with the keys "question" and "passage" to be consistent with the `reading_comprehension` predictor.
 
 ### Added